+++ decorator.rb
@@ -1,12 +1,12 @@
 require './nameable'
 
 class Decorator < Nameable
-    def initialize(nameable)
-        super()
-        @nameable = nameable
-    end
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
 
-    def correct_name
-        @nameable.correct_name
-    end
+  def correct_name
+    @nameable.correct_name
+  end
 end
