+++ nameable.rb
@@ -1,5 +1,5 @@
 class Nameable
-    def correct_name
-        raise NotImplementedError, "This method has not implemented yet."
-    end
+  def correct_name
+    raise NotImplementedError, 'This method has not been implemented yet.'
+  end
 end
