--- trimmerDecorator.rb
@@ -1,7 +1,0 @@
-require './decorator'
-
-class TrimmerDecorator < Decorator
-    def correct_name
-        @nameable.correct_name.slice(0, 10)
-    end
-end
